requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24       
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
et-xmlfile==1.1.0        
idna==3.4
joblib==1.2.0
nltk==3.7
openpyxl==3.0.10
Pillow==9.2.0
regex==2022.9.13
requests==2.28.1
sheets-tools==1.0.0      
soupsieve==2.3.2.post1   
tqdm==4.64.1
urllib3==1.26.12
waiting==1.4.1